Split up public operator-only tests
